protocol: make-only build system
This is	the protocol part of a make-only build system that
can normally coexist with the existing two build systems.

The primary benefit of the proposed approach is top-level build,
i.e. a way to make it easier to build all of SPICE with correct
dependency tracking when you make local changes in protocol or
common.	Last time I checked, it	was also faster	and easier to
maintain than the current alternatives