Server message to send dynamic adjustment of parameters to encoding agent
This message is sent by the server (typically in response to metrics
measured on the client indicating that the steam is presently not
optimal).

The message format is a tagged value, with a parameter ID and a 32-bit
parameter value, which will allow easy extensions in the future if new
useful parameters are found in codecs.

Parameters should only alter the quality of the stream, not the
overall functionality of the system.  As a result, it is always safe
to ignore unknown parameters.